refactor: move tx types to subdirectory
