+++ ethers-core/src/types/mod.rs
@@ -8,7 +8,8 @@ pub use ethabi::ethereum_types::H256 as TxHash;
 pub use ethabi::ethereum_types::{Address, Bloom, H160, H256, U128, U256, U64};
 
 mod transaction;
-pub use transaction::{Transaction, TransactionReceipt, TransactionRequest};
+pub use transaction::request::TransactionRequest;
+pub use transaction::response::{Transaction, TransactionReceipt};
 
 mod address_or_bytes;
 pub use address_or_bytes::AddressOrBytes;
